# backend/auth/api/http/openapi.gen.go
// Package http provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package http

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	strictecho "github.com/oapi-codegen/runtime/strictmiddleware/echo"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// EmailVerifyRequest defines model for EmailVerifyRequest.
type EmailVerifyRequest struct {
	// Token メール確認トークン（平文。INF-06はSHA-256ハッシュのみ保存し照合する）
	Token string `json:"token"`
}

// EmailVerifyResendRequest defines model for EmailVerifyResendRequest.
type EmailVerifyResendRequest struct {
	// Email VAR-01（RFC5322準拠・最大254文字）
	Email openapi_types.Email `json:"email"`
}

// LoginRequest defines model for LoginRequest.
type LoginRequest struct {
	// Email VAR-01（RFC5322準拠・最大254文字）
	Email openapi_types.Email `json:"email"`

	// Password 認証用パスワード（照合はbcrypt・CND-02）
	Password string `json:"password"`
}

// LoginResponse defines model for LoginResponse.
type LoginResponse struct {
	// AccessToken JWT（RS256・INF-03）。sub/roles/exp（VAR-03=15分）クレームを含む
	AccessToken string `json:"access_token"`

	// RefreshToken opaque token（INF-04・NFR-14。base64url。DBにはSHA-256ハッシュのみ保存）
	RefreshToken string `json:"refresh_token"`
}

// Problem RFC 9457 Problem Details（NFR-06）。retry_after・revocation_reason は拡張フィールド
type Problem struct {
	Detail *string `json:"detail,omitempty"`

	// ErrorCode 拡張エラーコード（VAR-11＝account_locked 等）
	ErrorCode *string `json:"error_code,omitempty"`
	Instance  *string `json:"instance,omitempty"`

	// RetryAfter レート制限・ロックアウト超過時の再試行可能秒数（E4・VAR-11）
	RetryAfter *int `json:"retry_after,omitempty"`

	// RevocationReason セッション失効理由コード（VAR-10・本APIでは未使用）
	RevocationReason *string `json:"revocation_reason,omitempty"`
	Status           int     `json:"status"`
	Title            string  `json:"title"`

	// Type VAR-15のベースURI配下（例 https://example.com/probs/validation-error）
	Type string `json:"type"`
}

// RegisterAccountRequest defines model for RegisterAccountRequest.
type RegisterAccountRequest struct {
	// Email VAR-01（RFC5322準拠・最大254文字）
	Email openapi_types.Email `json:"email"`

	// Password VAR-02（最小15文字・最大64文字・文字種の混在強制なし）
	Password string `json:"password"`
}

// VerifyEmailJSONRequestBody defines body for VerifyEmail for application/json ContentType.
type VerifyEmailJSONRequestBody = EmailVerifyRequest

// ResendEmailVerificationJSONRequestBody defines body for ResendEmailVerification for application/json ContentType.
type ResendEmailVerificationJSONRequestBody = EmailVerifyResendRequest

// LoginJSONRequestBody defines body for Login for application/json ContentType.
type LoginJSONRequestBody = LoginRequest

// RegisterAccountJSONRequestBody defines body for RegisterAccount for application/json ContentType.
type RegisterAccountJSONRequestBody = RegisterAccountRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// メールアドレスを確認する
	// (POST /auth/email-verify)
	VerifyEmail(ctx echo.Context) error
	// メール確認トークンを再送する
	// (POST /auth/email-verify/resend)
	ResendEmailVerification(ctx echo.Context) error
	// ログインする
	// (POST /auth/login)
	Login(ctx echo.Context) error
	// アカウントを登録する
	// (POST /auth/register)
	RegisterAccount(ctx echo.Context) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// VerifyEmail converts echo context to params.
func (w *ServerInterfaceWrapper) VerifyEmail(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.VerifyEmail(ctx)
	return err
}

// ResendEmailVerification converts echo context to params.
func (w *ServerInterfaceWrapper) ResendEmailVerification(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.ResendEmailVerification(ctx)
	return err
}

// Login converts echo context to params.
func (w *ServerInterfaceWrapper) Login(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.Login(ctx)
	return err
}

// RegisterAccount converts echo context to params.
func (w *ServerInterfaceWrapper) RegisterAccount(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.RegisterAccount(ctx)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(baseURL+"/auth/email-verify", wrapper.VerifyEmail)
	router.POST(baseURL+"/auth/email-verify/resend", wrapper.ResendEmailVerification)
	router.POST(baseURL+"/auth/login", wrapper.Login)
	router.POST(baseURL+"/auth/register", wrapper.RegisterAccount)

}

type VerifyEmailRequestObject struct {
	Body *VerifyEmailJSONRequestBody
}

type VerifyEmailResponseObject interface {
	VisitVerifyEmailResponse(w http.ResponseWriter) error
}

type VerifyEmail200Response struct {
}

func (response VerifyEmail200Response) VisitVerifyEmailResponse(w http.ResponseWriter) error {
	w.WriteHeader(200)
	return nil
}

type VerifyEmail400ApplicationProblemPlusJSONResponse Problem

func (response VerifyEmail400ApplicationProblemPlusJSONResponse) VisitVerifyEmailResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type VerifyEmail429ResponseHeaders struct {
	RetryAfter int
}

type VerifyEmail429ApplicationProblemPlusJSONResponse struct {
	Body    Problem
	Headers VerifyEmail429ResponseHeaders
}

func (response VerifyEmail429ApplicationProblemPlusJSONResponse) VisitVerifyEmailResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.Header().Set("Retry-After", fmt.Sprint(response.Headers.RetryAfter))
	w.WriteHeader(429)

	return json.NewEncoder(w).Encode(response.Body)
}

type ResendEmailVerificationRequestObject struct {
	Body *ResendEmailVerificationJSONRequestBody
}

type ResendEmailVerificationResponseObject interface {
	VisitResendEmailVerificationResponse(w http.ResponseWriter) error
}

type ResendEmailVerification200Response struct {
}

func (response ResendEmailVerification200Response) VisitResendEmailVerificationResponse(w http.ResponseWriter) error {
	w.WriteHeader(200)
	return nil
}

type ResendEmailVerification429ResponseHeaders struct {
	RetryAfter int
}

type ResendEmailVerification429ApplicationProblemPlusJSONResponse struct {
	Body    Problem
	Headers ResendEmailVerification429ResponseHeaders
}

func (response ResendEmailVerification429ApplicationProblemPlusJSONResponse) VisitResendEmailVerificationResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.Header().Set("Retry-After", fmt.Sprint(response.Headers.RetryAfter))
	w.WriteHeader(429)

	return json.NewEncoder(w).Encode(response.Body)
}

type ResendEmailVerification503ApplicationProblemPlusJSONResponse Problem

func (response ResendEmailVerification503ApplicationProblemPlusJSONResponse) VisitResendEmailVerificationResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(503)

	return json.NewEncoder(w).Encode(response)
}

type LoginRequestObject struct {
	Body *LoginJSONRequestBody
}

type LoginResponseObject interface {
	VisitLoginResponse(w http.ResponseWriter) error
}

type Login200JSONResponse LoginResponse

func (response Login200JSONResponse) VisitLoginResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type Login400ApplicationProblemPlusJSONResponse Problem

func (response Login400ApplicationProblemPlusJSONResponse) VisitLoginResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type Login401ApplicationProblemPlusJSONResponse Problem

func (response Login401ApplicationProblemPlusJSONResponse) VisitLoginResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(401)

	return json.NewEncoder(w).Encode(response)
}

type Login403ApplicationProblemPlusJSONResponse Problem

func (response Login403ApplicationProblemPlusJSONResponse) VisitLoginResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(403)

	return json.NewEncoder(w).Encode(response)
}

type Login429ResponseHeaders struct {
	RetryAfter int
}

type Login429ApplicationProblemPlusJSONResponse struct {
	Body    Problem
	Headers Login429ResponseHeaders
}

func (response Login429ApplicationProblemPlusJSONResponse) VisitLoginResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.Header().Set("Retry-After", fmt.Sprint(response.Headers.RetryAfter))
	w.WriteHeader(429)

	return json.NewEncoder(w).Encode(response.Body)
}

type RegisterAccountRequestObject struct {
	Body *RegisterAccountJSONRequestBody
}

type RegisterAccountResponseObject interface {
	VisitRegisterAccountResponse(w http.ResponseWriter) error
}

type RegisterAccount201Response struct {
}

func (response RegisterAccount201Response) VisitRegisterAccountResponse(w http.ResponseWriter) error {
	w.WriteHeader(201)
	return nil
}

type RegisterAccount400ApplicationProblemPlusJSONResponse Problem

func (response RegisterAccount400ApplicationProblemPlusJSONResponse) VisitRegisterAccountResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type RegisterAccount429ResponseHeaders struct {
	RetryAfter int
}

type RegisterAccount429ApplicationProblemPlusJSONResponse struct {
	Body    Problem
	Headers RegisterAccount429ResponseHeaders
}

func (response RegisterAccount429ApplicationProblemPlusJSONResponse) VisitRegisterAccountResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.Header().Set("Retry-After", fmt.Sprint(response.Headers.RetryAfter))
	w.WriteHeader(429)

	return json.NewEncoder(w).Encode(response.Body)
}

type RegisterAccount503ApplicationProblemPlusJSONResponse Problem

func (response RegisterAccount503ApplicationProblemPlusJSONResponse) VisitRegisterAccountResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(503)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// メールアドレスを確認する
	// (POST /auth/email-verify)
	VerifyEmail(ctx context.Context, request VerifyEmailRequestObject) (VerifyEmailResponseObject, error)
	// メール確認トークンを再送する
	// (POST /auth/email-verify/resend)
	ResendEmailVerification(ctx context.Context, request ResendEmailVerificationRequestObject) (ResendEmailVerificationResponseObject, error)
	// ログインする
	// (POST /auth/login)
	Login(ctx context.Context, request LoginRequestObject) (LoginResponseObject, error)
	// アカウントを登録する
	// (POST /auth/register)
	RegisterAccount(ctx context.Context, request RegisterAccountRequestObject) (RegisterAccountResponseObject, error)
}

type StrictHandlerFunc = strictecho.StrictEchoHandlerFunc
type StrictMiddlewareFunc = strictecho.StrictEchoMiddlewareFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// VerifyEmail operation middleware
func (sh *strictHandler) VerifyEmail(ctx echo.Context) error {
	var request VerifyEmailRequestObject

	var body VerifyEmailJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.VerifyEmail(ctx.Request().Context(), request.(VerifyEmailRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "VerifyEmail")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(VerifyEmailResponseObject); ok {
		return validResponse.VisitVerifyEmailResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// ResendEmailVerification operation middleware
func (sh *strictHandler) ResendEmailVerification(ctx echo.Context) error {
	var request ResendEmailVerificationRequestObject

	var body ResendEmailVerificationJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.ResendEmailVerification(ctx.Request().Context(), request.(ResendEmailVerificationRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "ResendEmailVerification")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(ResendEmailVerificationResponseObject); ok {
		return validResponse.VisitResendEmailVerificationResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// Login operation middleware
func (sh *strictHandler) Login(ctx echo.Context) error {
	var request LoginRequestObject

	var body LoginJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.Login(ctx.Request().Context(), request.(LoginRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "Login")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(LoginResponseObject); ok {
		return validResponse.VisitLoginResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// RegisterAccount operation middleware
func (sh *strictHandler) RegisterAccount(ctx echo.Context) error {
	var request RegisterAccountRequestObject

	var body RegisterAccountJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.RegisterAccount(ctx.Request().Context(), request.(RegisterAccountRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "RegisterAccount")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(RegisterAccountResponseObject); ok {
		return validResponse.VisitRegisterAccountResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}