# backend/auth/api/http/openapi.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Auth
  description: |
    auth-service の公開API。エラーレスポンスは RFC 9457（NFR-06・application/problem+json）。
paths:
  /auth/register:
    post:
      summary: アカウントを登録する
      operationId: registerAccount
      tags:
        - commands
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterAccountRequest'
      responses:
        '201':
          description: 登録を受け付けた（FR-01・メールアドレス重複時も同一応答＝列挙攻撃対策）
        '400':
          description: バリデーションエラー（VAR-01/VAR-02違反・E1/E2）
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://example.com/probs/validation-error
                title: Bad Request
                status: 400
        '429':
          description: 登録レートリミット超過（VAR-16・E4。登録済み/未登録に関わらず一様挙動。retry_afterはTTL残秒数）
          headers:
            Retry-After:
              description: 再試行までの秒数（bodyのretry_afterと同値）
              schema:
                type: integer
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://example.com/probs/rate-limit-exceeded
                title: Too Many Requests
                status: 429
                retry_after: 287
        '503':
          description: メール送信失敗（E3。登録・トークン保存はロールバック済み）
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://example.com/probs/mail-delivery-error
                title: Service Unavailable
                status: 503
  /auth/login:
    post:
      summary: ログインする
      operationId: login
      tags:
        - commands
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 認証成功（UC-005・STM-02.認証済み。アクセストークン〔JWT RS256〕とリフレッシュトークンを返す）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: メールアドレス形式不正（E1=validation-error）
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://example.com/probs/validation-error
                title: Bad Request
                status: 400
        '401':
          description: 認証失敗（E3=ユーザー不存在・E4=パスワード不一致。列挙防止のため同一応答）
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://example.com/probs/authentication-failed
                title: Unauthorized
                status: 401
        '403':
          description: メール未確認（E5=email-not-verified）または無効化済み（E6=account-disabled）
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://example.com/probs/email-not-verified
                title: Forbidden
                status: 403
        '429':
          description: ロックアウト中（E2・NFR-03。10回失敗で15分・retry_afterは解除までの秒数・VAR-11）
          headers:
            Retry-After:
              description: 解除までの秒数（bodyのretry_afterと同値）
              schema:
                type: integer
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://example.com/probs/account-locked
                title: Too Many Requests
                status: 429
                error_code: account_locked
                retry_after: 780
  /auth/email-verify:
    post:
      summary: メールアドレスを確認する
      operationId: verifyEmail
      tags:
        - commands
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailVerifyRequest'
      responses:
        '200':
          description: メールアドレスを確認した（UC-003・STM-01 mail_unverified→inactive・トークンは使い切り）
        '400':
          description: 無効・使用済みトークン（E1a/E1b=invalid-token・状態を区別しない）または期限切れ（E2=token-expired・CND-09不成立）
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://example.com/probs/invalid-token
                title: Bad Request
                status: 400
        '429':
          description: メール確認レートリミット超過（VAR-17・E4。IP単位・retry_afterはTTL残秒数）
          headers:
            Retry-After:
              description: 再試行までの秒数（bodyのretry_afterと同値）
              schema:
                type: integer
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://example.com/probs/rate-limit-exceeded
                title: Too Many Requests
                status: 429
                retry_after: 42
  /auth/email-verify/resend:
    post:
      summary: メール確認トークンを再送する
      operationId: resendEmailVerification
      tags:
        - commands
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailVerifyResendRequest'
      responses:
        '200':
          description: 受け付けた（UC-004・FR-04＝未登録/確認済みでも一律200・メール送信なし）。STM-01は不変（mail_unverifiedのまま）
        '429':
          description: 再送信レートリミット超過（VAR-13・E1。メールキー・一様適用・retry_afterはTTL残秒数）
          headers:
            Retry-After:
              description: 再試行までの秒数（bodyのretry_afterと同値）
              schema:
                type: integer
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://example.com/probs/rate-limit-exceeded
                title: Too Many Requests
                status: 429
                retry_after: 240
        '503':
          description: メール送信失敗（E2。新トークン発行・既存無効化はロールバック済み。ただしレート窓は消費）
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://example.com/probs/mail-delivery-error
                title: Service Unavailable
                status: 503
components:
  schemas:
    RegisterAccountRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          description: VAR-01（RFC5322準拠・最大254文字）
        password:
          type: string
          minLength: 15
          maxLength: 64
          description: VAR-02（最小15文字・最大64文字・文字種の混在強制なし）
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          description: VAR-01（RFC5322準拠・最大254文字）
        password:
          type: string
          description: 認証用パスワード（照合はbcrypt・CND-02）
    LoginResponse:
      type: object
      required:
        - access_token
        - refresh_token
      properties:
        access_token:
          type: string
          description: JWT（RS256・INF-03）。sub/roles/exp（VAR-03=15分）クレームを含む
        refresh_token:
          type: string
          description: opaque token（INF-04・NFR-14。base64url。DBにはSHA-256ハッシュのみ保存）
    EmailVerifyResendRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          description: VAR-01（RFC5322準拠・最大254文字）
    EmailVerifyRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          minLength: 1
          description: メール確認トークン（平文。INF-06はSHA-256ハッシュのみ保存し照合する）
    Problem:
      type: object
      description: RFC 9457 Problem Details（NFR-06）。retry_after・revocation_reason は拡張フィールド
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          description: VAR-15のベースURI配下（例 https://example.com/probs/validation-error）
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        retry_after:
          type: integer
          description: レート制限・ロックアウト超過時の再試行可能秒数（E4・VAR-11）
        error_code:
          type: string
          description: 拡張エラーコード（VAR-11＝account_locked 等）
        revocation_reason:
          type: string
          description: セッション失効理由コード（VAR-10・本APIでは未使用）